+++ src/main/kotlin/space/kscience/gradle/KScienceReadmeExtension.kt
@@ -7,21 +7,18 @@ import freemarker.template.TemplateNotFoundException
 import kotlinx.html.TagConsumer
 import kotlinx.html.div
 import kotlinx.html.stream.createHTML
-import kotlinx.validation.ApiValidationExtension
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.findByType
+import org.gradle.kotlin.dsl.*
 import org.intellij.lang.annotations.Language
+import org.jetbrains.dokka.gradle.AbstractDokkaTask
+import org.jetbrains.kotlin.gradle.dsl.abi.AbiValidationExtension
+import org.jetbrains.kotlin.gradle.dsl.abi.ExperimentalAbiValidation
+import space.kscience.gradle.internal.withKScience
+import space.kscience.gradle.internal.withKotlin
 import java.io.File
 import java.io.Serializable
 import java.io.StringWriter
 
-public enum class Maturity {
-    PROTOTYPE,
-    EXPERIMENTAL,
-    DEVELOPMENT,
-    STABLE,
-    DEPRECATED
-}
 
 private fun Template.processToString(args: Map<String, Any?>): String {
     val writer = StringWriter()
@@ -29,26 +26,13 @@ private fun Template.processToString(args: Map<String, Any?>): String {
     return writer.toString()
 }
 
+public class KScienceReadmeExtension(private val kscience: KSciencePlatformExtension) {
+    public val project: Project get() = kscience.project
 
-public class KScienceReadmeExtension(public val project: Project) {
     public var description: String? = null
         get() = field ?: project.description
 
-    public var maturity: Maturity = Maturity.EXPERIMENTAL
-        set(value) {
-            field = value
-            val projectName = project.name
-            if (value == Maturity.EXPERIMENTAL || value == Maturity.PROTOTYPE) {
-                project.rootProject.run {
-                    plugins.withId("org.jetbrains.kotlinx.binary-compatibility-validator") {
-                        extensions.findByType<ApiValidationExtension>()?.apply {
-                            project.logger.warn("$value project $projectName is excluded from API validation")
-                            ignoredProjects.add(projectName)
-                        }
-                    }
-                }
-            }
-        }
+    public var maturity: Maturity by kscience::maturity
 
     /**
      * If true, use default templates provided by plugin if override is not defined
@@ -86,7 +70,8 @@ public class KScienceReadmeExtension(public val project: Project) {
         templateLoader = fmLoader
     }
 
-    public data class Feature(val id: String, val description: String, val ref: String?, val name: String = id): Serializable
+    public data class Feature(val id: String, val description: String, val ref: String?, val name: String = id) :
+        Serializable
 
     public val features: MutableList<Feature> = mutableListOf()
 
@@ -192,7 +177,11 @@ public class KScienceReadmeExtension(public val project: Project) {
      */
     internal fun featuresString(itemPrefix: String = " - ", pathPrefix: String = ""): String = buildString {
         features.forEach {
-            appendLine("$itemPrefix[${it.name}]($pathPrefix${it.ref ?: "#"}) : ${it.description.lines().firstOrNull() ?: ""}")
+            appendLine(
+                "$itemPrefix[${it.name}]($pathPrefix${it.ref ?: "#"}) : ${
+                    it.description.lines().firstOrNull() ?: ""
+                }"
+            )
         }
     }
 
@@ -239,4 +228,72 @@ public class KScienceReadmeExtension(public val project: Project) {
     }
 
 
+}
+
+
+internal fun KSciencePlatformExtension.configureReadme() = with(project) {
+
+    //early return is readme is already configured
+    if (extensions.findByType<KScienceReadmeExtension>() != null) return@with
+
+    //Add readme generators to individual subprojects and root project
+    val readmeExtension = KScienceReadmeExtension(this@configureReadme)
+    this.extensions.add("readme", readmeExtension)
+
+    this@configureReadme.extensions.add("readme", readmeExtension)
+
+
+    val generateReadme by tasks.registering {
+        group = "documentation"
+        description = "Generate a README file if stub is present"
+
+        inputs.property("features", readmeExtension.features)
+
+        if (readmeExtension.readmeTemplate.exists()) {
+            inputs.file(readmeExtension.readmeTemplate)
+        }
+
+        readmeExtension.inputFiles.forEach {
+            if (it.exists()) {
+                inputs.file(it)
+            }
+        }
+
+        // add dependency for this task on subprojects readme tasks
+        subprojects {
+            withKScience {
+                extensions.findByType<KScienceReadmeExtension>()?.let { subProjectReadmeExtension ->
+                    tasks.findByName("generateReadme")?.let { readmeTask ->
+                        dependsOn(readmeTask)
+                    }
+                    inputs.property("features-${name}", subProjectReadmeExtension.features)
+                }
+            }
+        }
+
+        val readmeFile = file("README.md")
+        outputs.file(readmeFile)
+
+        doLast {
+            val readmeString = readmeExtension.readmeString()
+            if (readmeString != null) {
+                readmeFile.writeText(readmeString)
+            }
+        }
+    }
+
+    tasks.withType<AbstractDokkaTask> {
+        dependsOn(generateReadme)
+    }
+
+
+    // Enable API validation for production releases
+    if (!isInDevelopment && isMature()) {
+        withKotlin {
+            extensions.configure<AbiValidationExtension> {
+                @OptIn(ExperimentalAbiValidation::class)
+                enabled.set(true)
+            }
+        }
+    }
 }